requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
pydantic-settings
openai
google-genai
pydub
python-docx
sendgrid
PyMuPDF
python-pptx
aiofiles
httpx
